requirements.txt
requests==2.23.0
mysql-connector-python==8.0.19
